use hx711 lib instead custom impl
